services/kyc-service/main.py
from flask import Flask, request, jsonify
from datetime import datetime
import uuid
import random
import os
import psycopg2
from psycopg2.extras import RealDictCursor

app = Flask(__name__)

DATABASE_URL = os.getenv('DATABASE_URL', 'postgres://postgres:postgres@localhost:5432/baas')

def get_db():
    return psycopg2.connect(DATABASE_URL)

@app.route('/kyc/verify', methods=['POST'])
def verify_kyc():
    data = request.json
    user_id = data.get('user_id')
    document_type = data.get('document_type')
    document_number = data.get('document_number')
    face_image = data.get('face_image')
    document_image = data.get('document_image')
    
    # Simulação de IA para verificação (em produção, usar TensorFlow/PyTorch)
    face_match = round(0.92 + random.random() * 0.07, 4)
    liveness_score = round(0.88 + random.random() * 0.11, 4)
    aml_check = random.random() > 0.05
    
    status = 'approved' if face_match >= 0.85 and liveness_score >= 0.80 and aml_check else 'rejected'
    
    verification = {
        'id': str(uuid.uuid4()),
        'user_id': user_id,
        'document_type': document_type,
        'document_number': document_number,
        'face_match': face_match,
        'liveness_score': liveness_score,
        'aml_check': aml_check,
        'status': status,
        'verified_at': datetime.now().isoformat(),
        'created_at': datetime.now().isoformat()
    }
    
    return jsonify({
        'verification': verification,
        'message': 'KYC verification completed with AI',
        'details': {
            'face_match': face_match,
            'liveness_score': liveness_score,
            'aml_check': aml_check,
            'status': status,
            'processing_time_ms': 150
        }
    }), 200

@app.route('/kyc/aml-check', methods=['POST'])
def aml_check():
    data = request.json
    user_id = data.get('user_id')
    document_number = data.get('document_number')
    
    # Simulação de verificação AML com ML
    risk_score = round(random.random() * 0.3, 4)
    sanctions = random.random() > 0.95
    pep = random.random() > 0.97
    adverse_media = random.random() > 0.90
    
    status = 'flagged' if (sanctions or pep or adverse_media) else 'clear'
    
    return jsonify({
        'status': status,
        'risk_score': risk_score,
        'sanctions': sanctions,
        'pep': pep,
        'adverse_media': adverse_media,
        'checked_at': datetime.now().isoformat(),
        'data_sources': ['OFAC', 'UN', 'EU', 'Interpol', 'Local Authorities']
    }), 200

@app.route('/kyc/liveness', methods=['POST'])
def liveness_check():
    data = request.json
    video_data = data.get('video_data')
    
    # Simulação de detecção de deepfake com Deep Learning
    liveness_score = round(0.85 + random.random() * 0.14, 4)
    is_live = liveness_score > 0.80
    
    challenges = ['blink', 'smile', 'turn_head', 'nod']
    passed_challenges = [c for c in challenges if random.random() > 0.1]
    
    return jsonify({
        'is_live': is_live,
        'liveness_score': liveness_score,
        'challenges': challenges,
        'passed_challenges': passed_challenges,
        'deepfake_detected': not is_live,
        'deepfake_probability': round(1 - liveness_score, 4),
        'checked_at': datetime.now().isoformat(),
        'ai_model': 'DeepFake-Detector-v3.2'
    }), 200

@app.route('/kyc/document-ocr', methods=['POST'])
def document_ocr():
    data = request.json
    document_image = data.get('document_image')
    
    # Simulação de OCR com IA
    extracted_data = {
        'document_type': 'CPF',
        'document_number': '123.456.789-00',
        'full_name': 'João Silva Santos',
        'birth_date': '1990-05-15',
        'issue_date': '2020-01-10',
        'expiry_date': '2030-01-10',
        'confidence': round(0.95 + random.random() * 0.04, 4)
    }
    
    return jsonify({
        'extracted_data': extracted_data,
        'processing_time_ms': 200,
        'ocr_engine': 'Tesseract + Custom ML Model'
    }), 200

@app.route('/kyc/face-match', methods=['POST'])
def face_match():
    data = request.json
    face1 = data.get('face1')
    face2 = data.get('face2')
    
    # Simulação de comparação facial com Deep Learning
    similarity_score = round(0.88 + random.random() * 0.11, 4)
    is_match = similarity_score > 0.85
    
    return jsonify({
        'is_match': is_match,
        'similarity_score': similarity_score,
        'threshold': 0.85,
        'confidence': round(0.92 + random.random() * 0.07, 4),
        'ai_model': 'FaceNet + ArcFace',
        'processing_time_ms': 120
    }), 200

@app.route('/health', methods=['GET'])
def health():
    return jsonify({
        'status': 'healthy',
        'service': 'kyc-service',
        'time': datetime.now().isoformat(),
        'ai_models_loaded': True
    }), 200

if __name__ == '__main__':
    port = int(os.getenv('PORT', 8085))
    app.run(host='0.0.0.0', port=port, debug=False)
