services/kyc-service/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1
numpy==1.26.2
opencv-python==4.8.1.78
tensorflow==2.15.0
torch==2.1.2
face-recognition==1.3.0
pytesseract==0.3.10
Pillow==10.1.0
scikit-learn==1.3.2
